requirements.txt
accelerate==0.13.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diffusers==0.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
executing==1.1.1
fairscale==0.4.12
fastapi==0.85.1
fastjsonschema==2.16.2
filelock==3.8.0
firebase-admin==6.0.1
ftfy==6.1.1
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
httplib2==0.21.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
jedi==0.18.1
Jinja2==3.1.2
jose==1.0.0
jsonschema==4.16.0
jwcrypto==1.4.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
Pyrebase4==4.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.5
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.9.3
Send2Trash==1.8.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.20.4
terminado==0.16.0
timm==0.6.11
tinycss2==1.2.1
tokenizers==0.12.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.21.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.10.0